Introduce `ContextView`
Previously we needed to update the `environmentObject` calls in all
previews, which was error prone and tedious. Using the `ContextView`
allows us to make updates in one place, and it will be easier to
refactor the model layer if we want to pass models explicitly instead of
using `environmentObject` in the future.