Fix environment mirroring
We were applying the mirroring transform in `makeUIView` instead of
`updateUIView`, so it was not being applied when updating the facing
mode before the view is created. We currently get a mirrored frame of
the original facing mode before it changes, which we will fix in a
subsequent change.

Fixes #14.